pyproject.toml
[project]
name = "book-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "python-dotenv>=1.1.0",
    "social-auth-app-django>=5.4.3",
]
